Bug 11: Fix 'objective completed' false positive + add mission_test command (A371)
Root cause: Encounter objective type was 'agent' (report to agent), which the
EVE client auto-resolves as completed when player is docked at agent station.
Fix: Encounter objectives now use PyNone — dungeon section provides the
objective info (warp location, NPC kill tracking, completion status).

Also adds:
- 'mission_test' AI command queue command for server-side lifecycle testing
- test_combat_mission.py script to drive tests without client login
- EnsureEncounterBubble integration in mission_test for restart recovery

Test results: mission_test PASS — 5/5 NPCs spawned, IsMissionComplete=FALSE

+++ src/eve-common/utils/EvEMath.cpp
@@ -33,17 +33,17 @@ uint8 EvEMath::Skill::LevelForPoints(uint32 currentSP, uint8 rank)
     return (uint8)EvE::min(ret, EvESkill::MAXSKILLLEVEL);
 }
 
-uint8 EvEMath::Skill::PointsPerMinute(uint8 pAttr, uint8 sAttr)
+uint16 EvEMath::Skill::PointsPerMinute(uint8 pAttr, uint8 sAttr)
 {
-    return (pAttr + (0.5f * sAttr));
+    return (uint16)(pAttr + (0.5f * sAttr));
 }
 
-int64 EvEMath::Skill::StartTime(uint32 currentSP, uint32 nextSP, uint8 SPMin, int64 timeNow)
+int64 EvEMath::Skill::StartTime(uint32 currentSP, uint32 nextSP, uint16 SPMin, int64 timeNow)
 {
     return (timeNow - (((nextSP - currentSP) / SPMin) * EvE::Time::Minute));
 }
 
-int64 EvEMath::Skill::EndTime(uint32 currentSP, uint32 nextSP, uint8 SPMin, int64 timeNow)
+int64 EvEMath::Skill::EndTime(uint32 currentSP, uint32 nextSP, uint16 SPMin, int64 timeNow)
 {
     if (currentSP >= nextSP)
         return 0;
